requirements.txt
fastapi==0.95.0
PyJWT==2.6.0
pytest==7.2.2
uvicorn==0.21.1
httpx==0.23.3
pytest-asyncio==0.21.0
pytest-cov==4.0.0
